requirements.txt
# Core
openai>=1.30.0
anthropic>=0.39.0
mcp>=1.4.0
pydantic>=2.9.0
pyyaml>=6.0.2
python-dotenv>=1.0.1

# Account automation
easytrader>=0.24.0

# Data
akshare>=1.15.0
futu-api>=9.1.0

# Computation
numpy>=2.1.0
pandas>=2.2.0
ta>=0.11.0

# Output
tabulate>=0.9.0
rich>=13.9.0
